Move stuff around. Add tests.
