+++ packages/rut/src/mocks/event.ts
@@ -1,7 +1,6 @@
-/* eslint-disable complexity, max-classes-per-file */
-
 import React from 'react';
 import { EventType, EventOptions, InferHostElementFromEvent } from '../types';
+import { BaseEvent, SyntheticEvent, createHostEvent } from '../internals/event';
 
 /**
  * In both the `mockEvent` and `mockSyntheticEvent` functions below,
@@ -12,158 +11,11 @@ import { EventType, EventOptions, InferHostElementFromEvent } from '../types';
  * For example, if a test is dispatching `findOne('button').dispatch('onClick`)`,
  * the first argument is typed as `React.MouseEvent<HTMLButtonElement, MouseEvent>`.
  * We can take advantage of type inferrence by mocking the argument at
- * the call site, like so: `findOne('button').dispatch('onClick', {}, mockSyntheticEvent('onClick'))`.
+ * the call site, like so: `findOne('button').dispatch('onClick', mockSyntheticEvent('onClick'))`.
  *
  * With this pattern, mocks are easily typed, and the underlying event object
  * structure is close enough for most, if not all of test cases.
  */
-
-export class BaseEvent {
-  bubbles: boolean = true;
-
-  cancelable: boolean = true;
-
-  currentTarget: unknown = {};
-
-  defaultPrevented: boolean = false;
-
-  eventPhase: number = 0;
-
-  isTrusted: boolean = true;
-
-  propagationStopped: boolean = false;
-
-  srcElement: null = null;
-
-  target: unknown = {};
-
-  timeStamp: number;
-
-  type: string;
-
-  constructor(type: string) {
-    this.timeStamp = Date.now();
-    this.type = type;
-  }
-
-  preventDefault(): void {
-    this.defaultPrevented = true;
-  }
-
-  stopImmediatePropagation(): void {
-    this.propagationStopped = true;
-  }
-
-  stopPropagation(): void {
-    this.propagationStopped = true;
-  }
-}
-
-// https://developer.mozilla.org/en-US/docs/Web/Events
-// istanbul ignore next
-function createHostEvent(type: string): Event {
-  switch (type) {
-    case 'animationcancel':
-    case 'animationiteration':
-    case 'animationend':
-    case 'animationstart':
-      return new AnimationEvent(type);
-    case 'beforeunload':
-      return new BeforeUnloadEvent();
-    case 'auxclick':
-    case 'click':
-    case 'contextmenu':
-    case 'dblclick':
-    case 'mousedown':
-    case 'mouseenter':
-    case 'mouseleave':
-    case 'mousemove':
-    case 'mouseout':
-    case 'mouseover':
-    case 'mouseup':
-      return new MouseEvent(type);
-    case 'devicelight':
-    case 'devicemotion':
-    case 'deviceorientation':
-    case 'deviceorientationabsolute':
-      return new DeviceLightEvent(type);
-    case 'blur':
-    case 'focus':
-      return new FocusEvent(type);
-    case 'copy':
-    case 'cut':
-    case 'paste':
-      return new ClipboardEvent(type);
-    case 'compositionend':
-    case 'compositionstart':
-    case 'compositionupdate':
-      return new CompositionEvent(type);
-    case 'drag':
-    case 'dragend':
-    case 'dragenter':
-    case 'dragexit':
-    case 'dragleave':
-    case 'dragover':
-    case 'dragstart':
-    case 'drop':
-      return new DragEvent(type);
-    case 'error':
-      return new ErrorEvent(type);
-    case 'hashchange':
-      return new HashChangeEvent(type);
-    case 'keydown':
-    case 'keypress':
-    case 'keyup':
-      return new KeyboardEvent(type);
-    case 'loadend':
-    case 'progress':
-    case 'readystatechange':
-      return new ProgressEvent(type);
-    case 'message':
-    case 'messageerror':
-      return new MessageEvent(type);
-    case 'pagehide':
-    case 'pageshow':
-      return new PageTransitionEvent();
-    case 'gotpointercapture':
-    case 'lostpointercapture':
-    case 'pointercancel':
-    case 'pointerdown':
-    case 'pointerenter':
-    case 'pointerleave':
-    case 'pointerover':
-    case 'pointermove':
-    case 'pointerout':
-    case 'pointerup':
-      return new PointerEvent(type);
-    case 'popstate':
-      return new PopStateEvent(type);
-    case 'touchcancel':
-    case 'touchend':
-    case 'touchmove':
-    case 'touchstart':
-      return new TouchEvent(type);
-    case 'transitioncancel':
-    case 'transitionend':
-    case 'transitionrun':
-    case 'transitionstart':
-      return new TransitionEvent(type);
-    case 'abort':
-    case 'scroll':
-    case 'resize':
-      return new UIEvent(type);
-    case 'securitypolicyviolation':
-      return new SecurityPolicyViolationEvent(type);
-    case 'storage':
-      return new StorageEvent(type);
-    case 'unhandledrejection':
-      return new PromiseRejectionEvent(type, { promise: Promise.resolve() });
-    case 'wheel':
-      return new WheelEvent(type);
-    default:
-      return new Event(type);
-  }
-}
 
 /**
  * Mock a DOM `Event` based on type.
@@ -206,54 +58,6 @@ export function mockEvent<T = Event>(type: string, options?: EventOptions<Elemen
   return event;
 }
 
-export class SyntheticEvent extends BaseEvent {
-  nativeEvent: Event;
-
-  persisted: boolean = false;
-
-  constructor(type: string, nativeEvent: Event) {
-    super(type);
-
-    this.nativeEvent = nativeEvent;
-
-    // Copy over non-function properties
-    Object.entries(nativeEvent).forEach(([key, value]) => {
-      if (typeof value !== 'function') {
-        Object.defineProperty(this, key, {
-          enumerable: true,
-          value,
-        });
-      }
-    });
-  }
-
-  isDefaultPrevented(): boolean {
-    return this.defaultPrevented;
-  }
-
-  isPersistent(): boolean {
-    return this.persisted;
-  }
-
-  isPropagationStopped(): boolean {
-    return this.propagationStopped;
-  }
-
-  persist(): void {
-    this.persisted = true;
-  }
-
-  preventDefault(): void {
-    this.nativeEvent.preventDefault();
-    this.defaultPrevented = true;
-  }
-
-  stopPropagation(): void {
-    this.nativeEvent.stopPropagation();
-    this.propagationStopped = true;
-  }
-}
-
 /**
  * Mock a React `SyntheticEvent` based on type.
  */
@@ -269,4 +73,33 @@ export function mockSyntheticEvent<T = React.SyntheticEvent>(
 
   // @ts-ignore
   return new SyntheticEvent(eventType, mockEvent(eventType, options));
+}
+
+/**
+ * Factory a synthetic event from an an event directly, or with an options object.
+ */
+export function factorySyntheticEvent(
+  eventType: EventType,
+  event?: unknown,
+  elementType?: React.ElementType,
+): React.SyntheticEvent {
+  // Event provided by consumer, so use as is
+  if (event instanceof SyntheticEvent) {
+    return event as React.SyntheticEvent;
+  }
+
+  // Either event options or nothing provided
+  const options: EventOptions<Element, Event> =
+    typeof event === 'object' && event !== null ? event : {};
+
+  // Set a target automatically if not provided
+  if (typeof elementType === 'string' && !options.target) {
+    if (typeof document === 'undefined') {
+      options.target = { tagName: elementType.toUpperCase() };
+    } else {
+      options.target = document.createElement(elementType);
+    }
+  }
+
+  return mockSyntheticEvent(eventType, options);
 }
