+++ packages/rut/src/mocks/index.ts
@@ -1,5 +1,2 @@
-import { mockEvent, mockSyntheticEvent } from './event';
-
+export * from './event';
 export * from './fetch';
-
-export { mockEvent, mockSyntheticEvent };
